orders_api/app/requirements.txt
fastapi
mangum
pydantic
python-jose
boto3